pyproject.toml
[tool.poetry]
name = "pelican-pandoc-reader"
version = "0.0.1"
description = "Pelican plugin to render Pandoc Markdown files"
authors = ["Nandakumar Chandrasekhar <navanitachora@gmail.com>"]
license = "AGPL-3.0"
readme = "README.md"
keywords = ["pelican", "plugin", "pandoc"]
repository = "https://github.com/pelican-plugins/pandoc-reader"
documentation = "https://docs.getpelican.com"
packages = [
    { include = "pelican" },
]

classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Framework :: Pelican",
    "Framework :: Pelican :: Plugins",
    "Intended Audience :: End Users/Desktop",
    "Operating System :: OS Independent",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

[tool.poetry.urls]
"Funding" = "https://donate.getpelican.com/"
"Issue Tracker" = "https://github.com/pelican-plugins/pandoc-reader/issues"

[tool.poetry.dependencies]
python = "^3.6"
pelican = "^4.5"
markdown = {version = "^3.2.2", optional = true}
pyyaml = "^5.3.1"

[tool.poetry.dev-dependencies]
black = {version = "^19.10b0", allow-prereleases = true}
flake8 = "^3.8"
flake8-black = "^0.1.0"
invoke = "^1.3"
isort = "^5.4"
livereload = "^2.6"
markdown = "^3.2.2"
pytest = "^6.0"
pytest-cov = "^2.7"
pytest-pythonpath = "^0.7.3"
pytest-sugar = "^0.9.4"
Werkzeug = "^1.0"
pyyaml = "^5.3.1"

[tool.poetry.extras]
markdown = ["markdown"]

[tool.autopub]
project-name = "Pandoc Reader"
git-username = "botpub"
git-email = "botpub@autopub.rocks"

[tool.isort]
# Maintain compatibility with Black
combine_as_imports = true
force_grid_wrap = 0
include_trailing_comma = true
line_length = 88
multi_line_output = 3

# Sort imports within their section independent of the import type
force_sort_within_sections = true

[build-system]
requires = ["poetry>=1.0"]
build-backend = "poetry.masonry.api"
